Added custom HTML Plugin
