+++ packages/webpack-config/src/plugins/index.ts
@@ -1,2 +1,3 @@
 export { default as ExpoDefinePlugin } from './ExpoDefinePlugin';
 export { default as ExpoProgressBarPlugin } from './ExpoProgressBarPlugin';
+export { default as ExpoHtmlWebpackPlugin } from './ExpoHtmlWebpackPlugin';
